+++ src/snowflake/snowpark/modin/plugin/_internal/ordered_dataframe.py
@@ -680,6 +680,11 @@ class OrderedDataFrame:
             dataframe_ref.cached_snowflake_quoted_identifiers_tuple = tuple(
                 new_column_identifiers
             )
+            _logger.log(
+                # log at level DEBUG - 1 because we make this update very often.
+                level=logging.DEBUG - 1,
+                msg=f"The Snowpark DataFrame in DataFrameReference with id={dataframe_ref._id} is updated",
+            )
 
         new_df = OrderedDataFrame(
             dataframe_ref,